project/backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.45
pydantic==2.12.5
python-dotenv==1.2.1
psycopg2-binary==2.9.11
